+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 Release notes
 ==============
 
+### Unreleased
+
+- Fixed `Async<_>` not taking into account `[<GraphQLType>]`, e.g. if returning a boxed `obj` value for union-returning
+  fields
+
 ### 0.1.2 (2024-09-23)
 
 - Fixed option unwrapping not applying to `Async<_>` fields